Moved brewery and itinerary components into their own folders, also created an itinerary card component to better structure my JSX and to give each one a unique key
