+++ Brewtinerary/client/src/components/Home.js
@@ -1,5 +1,6 @@
 import React, { Component } from 'react';
 import ItineraryManager from '../API/itineraryManager';
+import ItineraryCard from './Itineraries/ItineraryCard';
 import "./css/Home.css";
 
 class Home extends Component {
@@ -25,28 +26,13 @@ class Home extends Component {
     return (
       <>
         <h1>My Itineraries</h1>
-        <table className="itineraries__table">
-          <thead>
-            <tr className="itinerary__tableRow">
-              <th>Name</th>
-              <th>Date</th>
-              <th>Location</th>
-            </tr>
-          </thead>
-          <tbody className="itinerary__tableBody">
             {
               this.props.itineraries.map(itinerary =>{
                 return(
-                  <tr>
-                    <td>{itinerary.name}</td>
-                    <td>{itinerary.dateOfEvent}</td>
-                    <td>{`${itinerary.city}, ${itinerary.state}`}</td>
-                  </tr>
+                  <ItineraryCard {...this.props} itinerary={itinerary} key={itinerary.id}/>
                 )
               })
             }
-          </tbody>
-        </table>
         <input onChange={this.handleInputChange} id="textToFilterBy" name="textToFilterBy" type="text" placeholder="Filter the itineraries" />
         <button onClick={this.filterValues}>Submit</button>
       </>
