+++ Brewtinerary/client/src/components/Itineraries/ItineraryCard.js
@@ -1,0 +1,14 @@
+import React from 'react';
+
+function ItineraryCard(props) {
+
+    return (
+        <div className="itineraryCard__div">
+            <h2>{props.itinerary.name}</h2>
+            <p>{props.itinerary.dateOfEvent}</p>
+            <p>{`${props.itinerary.city}, ${props.itinerary.state}`}</p>
+        </div>
+    )
+}
+
+export default ItineraryCard;
